scripts/claw/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claw"
version = "0.1.0"
description = "CLI wrapper around openpyxl/fitz/Pillow/etc. that absorbs the glue."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "claude-claw" }]
dependencies = [
  "click>=8.1",
  "click-plugins>=1.1",
  "rich>=13",
  "pyyaml>=6",
]

[project.scripts]
claw = "claw.__main__:cli"

[project.optional-dependencies]
xlsx = ["openpyxl>=3.1"]
docx = ["python-docx>=1"]
pptx = ["python-pptx>=0.6"]
pdf = ["pymupdf>=1.23", "pypdf>=4", "pdfplumber>=0.10", "reportlab>=4", "qrcode>=7"]
img = ["pillow>=10"]
web = ["httpx>=0.27", "trafilatura>=1.9"]
html = ["beautifulsoup4>=4.12", "lxml>=5", "soupsieve>=2.5"]
xml = ["lxml>=5"]
pipeline = ["networkx>=3"]
all = [
  "openpyxl>=3.1", "python-docx>=1", "python-pptx>=0.6",
  "pymupdf>=1.23", "pypdf>=4", "pdfplumber>=0.10", "reportlab>=4", "qrcode>=7",
  "pillow>=10",
  "httpx>=0.27", "trafilatura>=1.9",
  "beautifulsoup4>=4.12", "lxml>=5", "soupsieve>=2.5",
  "networkx>=3",
]
dev = ["pytest>=8", "syrupy>=4", "ruff>=0.6"]

[tool.hatch.build.targets.wheel]
packages = ["src/claw"]

[tool.ruff]
line-length = 100
